+++ client/cmdhficlass.c
@@ -777,10 +777,10 @@ int CmdHFiClassReader_Dump(const char *Cmd) {
 	if (have_debit_key) memcpy(tag_data+(3*8),div_key,8);
 	if (have_credit_key) memcpy(tag_data+(4*8),c_div_key,8);
 	// print the dump
-	printf("CSN   |00| %02X %02X %02X %02X %02X %02X %02X %02X |\n",tag_data[0],tag_data[1],tag_data[2]
-		  ,tag_data[3],tag_data[4],tag_data[5],tag_data[6],tag_data[7]);
-	printIclassDumpContents(tag_data, 1, (gotBytes/8)-1, gotBytes-8);
-
+	printf("------+--+-------------------------+\n");
+	printf("CSN   |00| %s|\n",sprint_hex(tag_data, 8));
+	printIclassDumpContents(tag_data, 1, (gotBytes/8), gotBytes);
+	
 	if (filename[0] == 0){
 		snprintf(filename, FILE_PATH_SIZE,"iclass_tagdump-%02x%02x%02x%02x%02x%02x%02x%02x",
 		    tag_data[0],tag_data[1],tag_data[2],tag_data[3],
@@ -1255,7 +1255,6 @@ int CmdHFiClass_loclass(const char *Cmd) {
 }
 
 void printIclassDumpContents(uint8_t *iclass_dump, uint8_t startblock, uint8_t endblock, size_t filesize) {
-	uint8_t blockdata[8];
 	uint8_t mem_config;
 	memcpy(&mem_config, iclass_dump + 13,1);
 	uint8_t maxmemcount;
@@ -1270,18 +1269,19 @@ void printIclassDumpContents(uint8_t *iclass_dump, uint8_t startblock, uint8_t e
 		startblock = 6;
 	if ((endblock > maxmemcount) || (endblock == 0))
 		endblock = maxmemcount;
-	if (endblock > filemaxblock)
+
+	// remember endblock need to relate to zero-index arrays.
+	if (endblock > filemaxblock-1)
 		endblock = filemaxblock;
+	
 	int i = startblock;
-	int j;
-	while (i <= endblock){
-		printf("Block |%02X| ",i);
-		memcpy(blockdata,iclass_dump + (i * 8),8);
-		for (j = 0;j < 8;j++)
-			printf("%02X ",blockdata[j]);
-		printf("|\n");
+	printf("------+--+-------------------------+\n");
+	while (i <= endblock) {
+		uint8_t *blk = iclass_dump + (i * 8);
+		printf("Block |%02X| %s|\n", i, sprint_hex(blk, 8) );	
 		i++;
 	}
+	printf("------+--+-------------------------+\n");
 }
 
 int usage_hf_iclass_readtagfile() {
@@ -1327,7 +1327,8 @@ int CmdHFiClassReadTagFile(const char *Cmd) {
 	size_t bytes_read = fread(dump, 1, fsize, f);
 	fclose(f);
 	uint8_t *csn = dump;
-	printf("CSN   [00] | %02X %02X %02X %02X %02X %02X %02X %02X |\n",csn[0],csn[1],csn[2],csn[3],csn[4],csn[5],csn[6],csn[7]);
+	printf("------+--+-------------------------+\n");
+	printf("CSN   |00| %s|\n", sprint_hex(csn, 8) );
 	//    printIclassDumpInfo(dump);
 	printIclassDumpContents(dump,startblock,endblock,bytes_read);
 	free(dump);